api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
requests